0x02 is unsigned. All Signed Ints are now Unsigned.
Preliminary testing has shown for this to connect and operate like normal. More testing is needed to make it gold.

* Log -1 is now 0, I don't know if that impacts.

+++ src/PFire.Core/Protocol/XFireAttributes/MessageAttribute.cs
@@ -9,13 +9,13 @@ namespace PFire.Core.Protocol.XFireAttributes
 {
     public class MessageAttribute : XFireAttribute
     {
-        private static readonly Dictionary<int, IMessage> MESSAGE_TYPES = new Dictionary<int, IMessage>
+        private static readonly Dictionary<uint, IMessage> MESSAGE_TYPES = new Dictionary<uint, IMessage>
         {
             { 0, new ChatMessage() },
             { 1, new ChatAcknowledgement() }
         };
 
-        private IMessage CreateMessage(short type)
+        private IMessage CreateMessage(uint type)
         {
             if(!MESSAGE_TYPES.ContainsKey(type))
             {
